modify to be a valid package
