+++ lib/dart_melty_soundfont.dart
@@ -1,0 +1,4 @@
+export 'src/synthesizer.dart';
+export 'src/synthesizer_settings.dart';
+export 'src/audio_renderer_ex.dart';
+export 'src/array_int16.dart';
