+++ lib/src/array_int16.dart
