+++ lib/src/array_math.dart
