+++ lib/src/binary_reader.dart
