+++ lib/src/channel.dart
