+++ lib/src/generator.dart
