+++ lib/src/generator_type.dart
