+++ lib/src/instrument_info.dart
