+++ lib/src/instrument_region.dart
