+++ lib/src/lfo.dart
