+++ lib/src/loop_mode.dart
