+++ lib/src/modulator.dart
