+++ lib/src/oscillator.dart
