+++ lib/src/preset_info.dart
