+++ lib/src/region_pair.dart
