+++ lib/src/soundfont_info.dart
